initial commit with a working version
this is the initial commit with a working version of the writereadport
utility. This utility is a command line application which can open a
Serial Communication Port and write to it and read from it. It has
several operation modes depending on command line arguments. Currently
we have implements the -s (settings) command line option and the -c
(console mode) option.